refactor: Web-Quest 구조 변경
