+++ tracky-core/src/main/java/kernel360/trackycore/core/domain/entity/QuestEntity.java
@@ -1,4 +1,4 @@
-package kernel360.trackyweb.question.domain;
+package kernel360.trackycore.core.domain.entity;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
