+++ tracky-core/src/main/java/kernel360/trackycore/core/infrastructure/repository/QuestRepository.java
@@ -1,9 +1,9 @@
-package kernel360.trackyweb.question.infrastructure;
+package kernel360.trackycore.core.infrastructure.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import kernel360.trackyweb.question.domain.QuestEntity;
+import kernel360.trackycore.core.domain.entity.QuestEntity;
 
 @Repository
 public interface QuestRepository extends JpaRepository<QuestEntity, Long> {
